requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# AI/ML
google-generativeai==0.8.0
pytesseract==0.3.10
opencv-python==4.9.0.80
scikit-learn==1.4.0
faiss-cpu==1.7.4
nltk==3.8.1

# Dashboard
streamlit==1.31.0
streamlit-aggrid==0.3.4.1
plotly==5.18.0

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0
Pillow==10.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
